+++ package.json
@@ -7,6 +7,7 @@
     "build": "npm run build"
   },
   "dependencies": {
+    "async": "^3.2.6",
     "axios": "^1.7.4",
     "bcryptjs": "^2.4.3",
     "cookie-parser": "~1.4.4",
